requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
mysql-connector-python==8.2.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
prophet==1.1.5
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
joblib==1.3.2
openpyxl==3.1.2
Werkzeug==3.0.1